Add files via upload
Added a winner view and url. Added a button on the home page to perform the raffle. Draw button links to the winner page. The random library was used to pick a random winner. The winner variable is displayed on the winner page. A link to perform a new raffle was added to the winner page.